Fix 'cannot read property "iso"' crash (issue #80)
On visiting route that is not defined (404), and module's `defaultLocale`
is not defined (default), the code in main.js would set `app.i18n.locale`
to null and when the seo code would run, it would crash on trying to
access property `iso` of it. So don't allow seo code to run when locale
is not defined.