test: use @nuxtjs/module-test-utils testing helper
This cleans up tests a bit and also uses getPort package which makes
it possible to use random port for testing so that it doesn't conflict
with anything or tests itself.

Due to how utils work, moved test file to the root of the 'test'
directory instead of being inside fixtures directory.

Renamed 'fixtures' directory to 'fixture' as that is what utils expect.