fix: Locale prefixes missing for child routes with custom paths
If one assigned (in component options or pages object) custom, absolute
paths for routes and those routes were child routes, then those routes
didn't get the locale prefix assigned as prefix was never added to
child routes.

That wasn't a problem if route was a top-level route or no custom paths
were assigned as then the path was relative and didn't need locale
prefix to be added.

Fix by also adding locale prefix for routes that are child routes with
absolute path.

Resolves #359