fix: make switchLocalePath work from Nuxt plugin or middleware
switchLocalePath didn't work from a Nuxt plugin or middleware
(from context.app) as it accessed properties on `this` that were not
accessible from there.

Fixed by creating a proxy interface that is consistent for both
Vue instances and Nuxt context so that switchLocalePath can access
same properties from it and get expected results.

Also:
  - Updated `differentDomains` tests to not use snapshot testing.
    Those actually were wrongly tested due to bug in test utils but
    failure wasn't caught in snapshot (it was supposed to have
    SEO disabled).

Resolves #480