fix(routing): Restore handling of route argument in getRouteBaseName
In 8a1c052f340ad8f86d62fb1e3c6f1107360fb0e9 the functionality to pass
route to `getRouteBaseName` was accidentally removed. Bring that back
and update documentation to state what's the actual argument. It was
incorrectly specifying `string | RawLocation` as argument while
currently it has to be a `Route`.

Resolves #539