fix: crash with invalid locale cookie + no_prefix strategy + SEO (#666)
The SEO code has crashed due to trying to get `iso` property for locale
that didn't exist. It's because we've allowed to set `$i18n.locale` to
an invalid value due to not validating the locale cookie in all cases.

This crash only triggered when certain conditions have aligned:
 - using `no_prefix` strategy as then middleware can't get locale from route
 - using `detectBrowserLanguage.alwaysRedirect` as then we try to redirect
   even when cookie is already set
 - using `seo = true` or manual SEO integration as the crash happened
   in SEO code

Change `getLocaleCookie` to only return locale code if it's valid.