fix: remove "encodePaths" option and don't encode paths by default
This could be considered a breaking change but since I only know of it
breaking things rather than helping anything and there are no tests that
would prove me otherwise, I think it's safe to remove without releasing
a major version.

Also, Nuxt doesn't encode paths either so this change aligns with it.

Resolves #712