fix: don't add trailing slashes to generated routes
Module was adding a trailing slash for extra routes that
it was generating.

This isn't normally a problem as VueRouter runs by default in
non-strict mode and normalizes paths to not have trailing slashes but
when the "router.trailingSlash = false" option was set in Nuxt, then
the matching was strict and actually made trailing slashes appear in
the path.

Follow the same logic as Nuxt and:
 - don't add trailing slashes by default or when "router.trailingSlash"
   option is set to "false"
 - add trailing slashes when "router.trailingSlash" option is set
   to "true"

Resolves #717