Change return type of localeRouter from Route to Location
Event though the result is actually a route object, the point of
localeRoute is to return something that can be passed to VueRouter.push()
and that needs to be a Location object.

Location is basically a subset of Route plus some optional properties
so treating Route as Location should be fine.

Resolves #776