fix(lazy): make client use translations loaded on the server
Instead of client importing and processing messages (translations)
itself on load pass already loaded messages from the server through
nuxtState. This will in most cases allow avoiding triggering an extra
network request to fetch lang file for given locale.

Slight behavior change: The client will no longer import the lang file
itself for the initially used locale. That means that if the lang file
has exported a function, it will only be called on the server and not
the client. This matches behavior of "asyncData" and I think it makes
more sense in general.

Resolves #486
Resolves #663