fix(differentDomains): Match domain properly on client if port provided
If locale's "domain" contained a port then that wasn't matched correctly
on the client as it compared with "location.hostname" that doesn't include
the port. Use "location.host" instead.

With "vuex.syncLocale" set to to "true" the problem was not always
visible as server-determined locale would be used on the client.

Resolves #831