fix: don't resolve paths to non-existent routes
If the resolved path for a given locale is 404 then don't attempt to
change the path to locale specific. Avoids redirecting unnecessarily to
a route that doesn't exist anyway.

That also fixes the security issue with redirecting to a different
domain but just in case added an additional measure against that.

Resolves #1092