refactor(lazy): include fallback locale messages in the main bundle (#1129)
This is an optimization to avoid including the messages for the fallback
locale in the Nuxt state.

Since the messages for the fallback locale are **always required** when
"vueI18n.fallbackLocale" is defined we can include them in the main bundle
and avoid making all server-side responses include it separately. The
amount of data is the same in both cases but this way we can take
advantage of browser caching.